% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-pilr-api-function.r
\name{gen_pilr_api_function}
\alias{gen_pilr_api_function}
\title{Return function to use PiLR API.}
\usage{
gen_pilr_api_function(api_string)
}
\arguments{
\item{api_string}{The string to append to the API call string.}
}
\description{
This function will return a function suitable to querying the PiLR
API.
}
