% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_groups}
\alias{get_groups}
\title{Return projects from the PiLR server.}
\usage{
get_groups(project, pilr_server, access_code, ...)
}
\arguments{
\item{pilr_server}{the pilr_server to get the list of projects from}
}
\description{
This function will return a data.frame containing the names and
identification codes of projects the user has access to on the PiLR
server.
}
\examples{
get_projects("localhost:8080")
}
\keyword{projects}

