% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_participants}
\alias{get_participants}
\title{Return participants from the PiLR server.}
\usage{
get_participants(project, pilr_server, access_code, ...)
}
\arguments{
\item{project}{the PiLR project string identifier}

\item{pilr_server}{URL of PiLR server}
}
\description{
This function will return a data.frame containing the names and
identification codes of participants for a given project.
}
\examples{
get_projects("localhost:8080")
}
\keyword{participants}

