% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_projects}
\alias{get_projects}
\title{Return project information from the PiLR server.}
\usage{
get_projects(pilr_server, access_code)
}
\arguments{
\item{pilr_server}{a character vector of length 1 indicating the
PiLR server URL}
}
\description{
This function will return a data.frame containing the names and
unique identifiers of projects the user has access to on the PiLR
server.
}
\examples{
get_projects("localhost:8080")
}
\keyword{projects}

