% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_schemas}
\alias{get_schemas}
\title{Return data set schema information from the PiLR server.}
\usage{
get_schemas(project, data_set, pilr_server, access_code, ...)
}
\arguments{
\item{project}{the pilr_server to get the list of projects from}

\item{data_set}{the pilr_server to get the list of projects from}

\item{pilr_server}{a character vector of length 1 indicating the
pilr_server URL}
}
\description{
This function will return a data.frame containing the names and
identification codes of schemas for a data set.
}
\examples{
get_projects("localhost:8080")
}
\keyword{projects}

