% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_json_file}
\alias{post_json_file}
\title{Write data contained in a JSON file to the PiLR server.}
\usage{
post_json_file(file, endpoint, ...)
}
\arguments{
\item{pilr_server}{the pilr_server to get the list of projects from}
}
\description{
This function will return a data.frame containing the names and
identification codes of projects the user has access to on the PiLR
server.
}
\examples{
get_projects("localhost:8080")
}
\keyword{projects}

