% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pilr.r
\name{read_pilr}
\alias{read_pilr}
\title{Return measurement data from the PiLR server.}
\usage{
read_pilr(pilr_server, project, data_set, schema, access_code,
  bad_metadata_ok = FALSE, value = "parsed", allrecords = FALSE,
  query_params = NULL, ...)
}
\arguments{
\item{endpoint}{The endpoint to retrieve information.}
}
\description{
This function will return a data.frame containing measurement data
}
\examples{
read_pilr(endpoint)

}
\keyword{data}
\keyword{measurement}
